+++ Wedding/frontend/src/pages/AdminPage.js
@@ -71,6 +71,7 @@ const AdminPage = () => {
 						/>
 						<InviteeTable
 							invitees={invitees}
+							addInvitee={addInvitee}
 							updateInvitee={updateInvitee}
 							removeInvitee={removeInvitee}
 							selectedInvitees={selectedInvitees}